Some ui changes №6

+++ src/helpers/cars_data.cpp
@@ -4,9 +4,10 @@ CarsData::CarsData(std::vector<CarAchievements> new_cars_data)
     : cars_data(std::move(new_cars_data)) {
 }
 
-double CarsData::GetVelocity(int index) const {
+size_t CarsData::GetVelocity(int index) const {
   double current_velocity = cars_data[index].current_showed_velocity;
-  return current_velocity < kMinVisibleVelocity ? 0 : current_velocity;
+  return current_velocity < kMinVisibleVelocity ?
+         0 : static_cast<size_t>(current_velocity);
 }
 
 int32_t CarsData::GetLapsCounter(int index) const {
